Rename categories to todoFilters
Rename categories to todoFilters
