+++ src/constants/config.js
@@ -1,2 +1,7 @@
 ﻿const categoryAll = { id: '0', name: 'All' };
+
+export const ALL_TODOS = 'ALL_TODOS';
+export const ONLY_COMPLETED = 'ONLY_COMPLETED';
+export const ONLY_TO_COMPLETE = 'ONLY_TO_COMPLETE';
+
 export default categoryAll;
