+++ src/containers/TodosContainer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import Todos from '../components/Todos';
-import { fetchAllCategories } from '../actions/categoriesActions';
+import { fetchAllCategories } from '../actions/todoFiltersActions';
 import { hideMessage } from '../actions/messageActions';
 
 const TodosContainer = props => <Todos {...props} />;
